requirements.txt
beautifulsoup4==4.12.3
cvxpy==1.4.2
entmax==1.3
gurobipy==11.0.1
gymnasium==0.29.1
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
ortools==9.9.3963
pandas==2.2.1
pytorch_lightning==2.2.1
pyunpack==0.3
PyYAML==6.0.1
ray==2.9.3
Requests==2.31.0
scikit_learn==1.4.1.post1
scipy==1.12.0
setuptools==69.2.0
six==1.16.0
tqdm==4.66.2
PySCIPOpt==4.4.0
